Split image reference slot state
