+++ frontend/app/(core)/(workspace)/app/image/AGENTS.md
@@ -15,6 +15,7 @@ This route owns the dedicated still-image generation workspace.
 - Keep reusable image workspace types and storage constants in `_lib/image-workspace-types.ts`.
 - Keep character reference parsing, labels, and slot merging in `_lib/image-workspace-character-references.ts`.
 - Keep job-to-history and local active group builders in `_lib/image-workspace-history.ts`.
+- Keep reference slot state, upload orchestration, library selection, and character toggling in `_hooks/useImageReferenceSlots.ts`.
 - Keep browser-only library UI in `_components/ImageLibraryModal.tsx`.
 - Keep `ImageWorkspace.tsx` focused on orchestration; do not add new inline helper blocks or route-local modal implementations there.
 - Keep image composer storage keys and stored payload shape backward compatible.
